Add relation between images and experience-tree-node-score
This relationship will be used to attach images to the
tree-node-score.  The grand idea here is that the information used
from attaching the images to the most detailed
experience-tree-node-score elements would allow us to render the
images at specific places and in a specific order.  This will result
in some extra load on the whole system, but we suspect we'll be able
to cache the images.