Capture and present focus value of tree nodes for specific groups
Some properties are especially valuable to specific groups.  The
constructed tree can receive focus points in this way.  We now capture
and expose this information in a first approach.

The approach we take on this may change.  We may convert this into a
code-list and creating relationships to these code-lists.  That could be
a more flexible model for the future.